requirements.txt
mfrc522==0.0.7
mysql-connector-python==9.3.0
RPi.GPIO==0.7.1
spidev==3.6
